need to fix batchwise endpoint

+++ src/main/java/com/thbs/cpt/DTO/BatchWiseProgressDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 // @NoArgsConstructor
 public class BatchWiseProgressDTO {
-    private int batchId;
+    private long batchId;
     private double batchProgress;
 
 }
